Normalize AI agent uploads via shared image utils
