+++ src/components/profile/edit/overview/HeroRow.tsx
@@ -3,6 +3,8 @@
 import Image from "next/image";
 import React, { ChangeEvent, useRef, useState } from "react";
 
+import { isLikelyImage, normalizeImageToJpeg } from "@/helpers/utils/image";
+
 type Props = {
   userName?: string;
   avatarUrl?: string | null; // URL из бэка (если уже загружен)
@@ -11,120 +13,6 @@ type Props = {
   canRemoveAvatar?: boolean;
   description?: string;
 };
-
-const MAX_TARGET_LONG_SIDE = 1600; // px
-const JPEG_QUALITY = 0.85;
-
-// Разрешённые растровые форматы (SVG исключаем для аватаров)
-const RASTER_EXTS = [
-  "jpg",
-  "jpeg",
-  "png",
-  "webp",
-  "gif",
-  "heic",
-  "heif",
-  "bmp",
-  "tif",
-  "tiff",
-];
-
-function getFileExtLower(name: string) {
-  const i = name.lastIndexOf(".");
-  return i === -1 ? "" : name.slice(i + 1).toLowerCase();
-}
-
-function isLikelyImage(file: File): boolean {
-  // 1) Надёжнее всего по MIME
-  if (file.type && file.type.startsWith("image/")) return true;
-  // 2) Фоллбек по расширению (на iOS иногда type=="")
-  const ext = getFileExtLower(file.name);
-  return RASTER_EXTS.includes(ext);
-}
-
-/** Быстрый декодер (если поддерживается браузером) */
-async function readAsImageBitmap(file: File): Promise<ImageBitmap> {
-  return await createImageBitmap(file);
-}
-
-/** Канвас + ресайз с высоким качеством */
-function imageToCanvas(img: HTMLImageElement | ImageBitmap) {
-  let width: number;
-  let height: number;
-
-  if (img instanceof HTMLImageElement) {
-    width = img.naturalWidth;
-    height = img.naturalHeight;
-  } else {
-    width = img.width;
-    height = img.height;
-  }
-
-  const maxSide = Math.max(width, height);
-  const scale = maxSide > MAX_TARGET_LONG_SIDE ? MAX_TARGET_LONG_SIDE / maxSide : 1;
-  const targetW = Math.round(width * scale);
-  const targetH = Math.round(height * scale);
-
-  const canvas = document.createElement("canvas");
-  canvas.width = targetW;
-  canvas.height = targetH;
-  const ctx = canvas.getContext("2d");
-  if (!ctx) throw new Error("Canvas 2D context unavailable");
-  ctx.imageSmoothingEnabled = true;
-  ctx.imageSmoothingQuality = "high";
-
-  ctx.drawImage(img as CanvasImageSource, 0, 0, width, height, 0, 0, targetW, targetH);
-
-  return canvas;
-}
-
-/** Фоллбек-декод через <img> для почти всех форматов, включая HEIC в Safari */
-async function fileToHtmlImage(file: File): Promise<HTMLImageElement> {
-  const url = URL.createObjectURL(file);
-  const img = await new Promise<HTMLImageElement>((resolve, reject) => {
-    const el = document.createElement("img") as HTMLImageElement;
-    el.onload = () => resolve(el);
-    el.onerror = reject;
-    el.src = url;
-  });
-  return img;
-}
-
-/**
- * Нормализация входного файла:
- * - декодируем (ImageBitmap или <img>)
- * - уменьшаем до разумного размера
- * - конвертируем в JPEG (image/jpeg)
- */
-async function normalizeImageToJpeg(file: File): Promise<File> {
-  try {
-    const bitmap = await readAsImageBitmap(file);
-    const canvas = imageToCanvas(bitmap);
-    const blob: Blob = await new Promise((resolve) =>
-      canvas.toBlob((b) => resolve(b as Blob), "image/jpeg", JPEG_QUALITY)
-    );
-    return new File([blob], replaceExt(file.name, ".jpg"), {
-      type: "image/jpeg",
-      lastModified: Date.now(),
-    });
-  } catch {
-    const img = await fileToHtmlImage(file);
-    const canvas = imageToCanvas(img);
-    const blob: Blob = await new Promise((resolve) =>
-      canvas.toBlob((b) => resolve(b as Blob), "image/jpeg", JPEG_QUALITY)
-    );
-    return new File([blob], replaceExt(file.name, ".jpg"), {
-      type: "image/jpeg",
-      lastModified: Date.now(),
-    });
-  }
-}
-
-function replaceExt(name: string, ext: string) {
-  const i = name.lastIndexOf(".");
-  if (i === -1) return name + ext;
-  return name.slice(0, i) + ext;
-}
 
 export default function HeroRow({
   userName,
